// kaliber-platform-complete/kaliber-webapplication/kaliber-webapp/src/app/interfaces/quiz-inventory.ts
export interface Quiz {
    title: string;
    quizCode: string;
    createdBy: string;
    createDate: any;
    subject: string;
    quizInstructions: string;
    concepts: any;
    maxDurationMinutes: number;
    maxScore: any;
    description: string;
    status: any;
    visibility: any;
    proficiency: any;
    maxAttempts: any;
    maxQuestions: number;
   }